+++ rs/moq-mux/src/hang/mod.rs
@@ -1,7 +1,7 @@
 mod container;
 
-pub use container::Legacy;
+pub use container::{Legacy, Media};
 
-pub type Consumer = crate::ordered::Consumer<Legacy>;
-pub type Producer = crate::ordered::Producer<Legacy>;
+pub type Consumer = crate::ordered::Consumer<Media>;
+pub type Producer = crate::ordered::Producer<Media>;
 pub use crate::container::Frame;