+++ rs/moq-mux/src/lib.rs
@@ -1,14 +1,13 @@
 //! Media muxers and demuxers for MoQ.
 
 mod catalog;
-#[cfg(feature = "mp4")]
 pub mod cmaf;
 pub mod container;
 mod error;
 pub mod hang;
-pub mod import;
 pub mod msf;
 pub mod ordered;
+pub mod producer;
 
 pub use catalog::*;
 pub use error::*;